pyproject.toml
[project]
name = "family-expense-report"
version = "0.1.0"
description = ""
authors = [
    {name = "None"}
]
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "argparse (>=1.4.0,<2.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]

[tool.poetry]
name = "family-expense-report"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.13.5"
pyyaml = ">=6.0.2,<7.0.0"
pandas = ">=2.3.0,<3.0.0"
pathlib = ">=1.0.1,<2.0.0"
argparse = ">=1.4.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
